계좌 목록 조회 로직 추가

+++ src/main/java/emcast/subject/dto/service/AccountListResponse.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 @Getter
@@ -14,11 +14,11 @@ public class AccountListResponse {
 
     private String userName;
 
-    private BigDecimal totalBalance; // 전체 계좌를 합친 금액 (일반 + 적금)
+    private BigInteger totalBalance; // 전체 계좌를 합친 금액 (일반 + 적금)
 
-    private BigDecimal totalRegularBalance; // 전체 일반계좌를 합친 금액
+    private BigInteger totalRegularBalance; // 전체 일반계좌를 합친 금액
 
-    private BigDecimal totalSavingsBalance; // 전체 적금계좌를 합친 금액
+    private BigInteger totalSavingsBalance; // 전체 적금계좌를 합친 금액
 
     private List<AccountInfo> accountInfos;
 }
